Store/retrieve logs from an array instead of system log
I found that using NSLogv and retrieving logs from the system log was
unreliable; it omitted many of my logs and included some strange stuff,
and didn’t have dates.  (Also I didn’t like the idea of my logs getting
stored in system log). This commit makes the class store logs in a
singleton instance of a mutable array instead of the system log, with
date/time of log, capped at a certain number of logs. The log can be
saved in
applicationWillTerminate: so that it remains continuous through crashes
and force quits.